Integration of TestCafe and CucumberJS
This is a sample project which demonstrates the integration between Testcafe and cucumber. Here we used to test the login functionality with the help of feature file[Gherkin Syntax] and implemented with Testcafe code. For reporting we use "cucumber-html-reporter" which converts the result JSON to a bootstrap themed html report.

Note:
Make sure Node.js is installed
Navigate to the root of the repo
Use the npm install command
Check node_modules are created under the project